chore: modal stack 관련 작업 훅으로 분리
